Changes in index page
* registration-closed page was removed
* RegistrationClosedForm component was created
* Moved form from index to MainForm component
* The conditional was made in index for the rendering of the components
