+++ app/components/RegistrationClosedForm.js
